@@ -8,9 +8,9 @@
 
 	<properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
+        <maven-compiler-plugin.version>3.14.0</maven-compiler-plugin.version>
         <maven-resources-plugin.version>3.1.0</maven-resources-plugin.version>
-        <java.version>1.8</java.version>
+        <java.version>21</java.version>
         <maven.test.skip>true</maven.test.skip>
         <fastjson.version>1.2.70</fastjson.version>
         <jackson-databind.version>2.14.2</jackson-databind.version>
@@ -184,13 +184,13 @@
             <dependency>
                 <groupId>org.apache.shiro</groupId>
                 <artifactId>shiro-spring-boot-web-starter</artifactId>
-                <version>2.0.3</version>
+                <version>1.9.0</version>
             </dependency>
 
             <dependency>
                 <groupId>org.hibernate.validator</groupId>
                 <artifactId>hibernate-validator</artifactId>
-                <version>8.0.1.Final</version>
+                <version>6.0.15.Final</version>
             </dependency>
 
             <!-- MySQL 连接驱动依赖 -->
@@ -333,9 +333,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>${maven-compiler-plugin.version}</version>
                 <configuration>
-                    <source>${java.version}</source>
-                    <target>${java.version}</target>
                     <encoding>${project.build.sourceEncoding}</encoding>
+     <release>${java.version}</release>
                 </configuration>
             </plugin>
             <plugin>
@@ -355,14 +354,14 @@
                     <exportDatatables>true</exportDatatables>
                     <activeRecipes>
                         <recipe>org.openrewrite.java.migrate.UpgradeToJava21</recipe>
-                        <!--Spring 升级配置-->
-                        <recipe>com.yourorg.UpgradeExplicitSpringBootDependenciesExample</recipe>
-                        <recipe>org.openrewrite.java.spring.cloud2023.UpgradeSpringCloud_2023</recipe>
-                        <!--HibernateValidator 升级-->
-                        <recipe>org.openrewrite.hibernate.validator.HibernateValidator_8_0</recipe>
-                        <!--升级 Spring 需要升级 Jakarta 10  -->
-                        <recipe>org.openrewrite.java.migrate.jakarta.UpdateJakartaXmlWsEE10</recipe>
-                        <recipe>org.openrewrite.java.migrate.jakarta.UpdateApacheShiroDependencies</recipe>
+                        <!--&lt;!&ndash;Spring 升级配置&ndash;&gt;-->
+                        <!--<recipe>org.openrewrite.maven.spring.UpgradeExplicitSpringBootDependencies</recipe>-->
+                        <!--<recipe>org.openrewrite.java.spring.cloud2023.UpgradeSpringCloud_2023</recipe>-->
+                        <!--&lt;!&ndash;HibernateValidator 升级&ndash;&gt;-->
+                        <!--<recipe>org.openrewrite.hibernate.validator.HibernateValidator_8_0</recipe>-->
+                        <!--&lt;!&ndash;升级 Spring 需要升级 Jakarta 10  &ndash;&gt;-->
+                        <!--<recipe>org.openrewrite.java.migrate.jakarta.UpdateJakartaXmlWsEE10</recipe>-->
+                        <!--<recipe>org.openrewrite.java.migrate.jakarta.UpdateApacheShiroDependencies</recipe>-->
                     </activeRecipes>
                 </configuration>
                 <dependencies>
