@@ -1,12 +1,14 @@
 package com.haust.shop.admin.util;
 
+import java.io.Serial;
 import java.io.Serializable;
 
 import com.haust.shop.db.domain.DtsBrand;
 
 public class DtsBrandVo extends DtsBrand implements Serializable{
 
-	private static final long serialVersionUID = 6530090986580196500L;
+    @Serial
+    private static final long serialVersionUID = 6530090986580196500L;
 	
 	private Integer[] categoryIds;
 
