@@ -8,7 +8,6 @@ import java.io.InputStream;
 import java.net.MalformedURLException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.stream.Stream;
 
@@ -29,7 +28,7 @@ public class LocalStorage implements Storage {
 	public void setStoragePath(String storagePath) {
 		this.storagePath = storagePath;
 
-		this.rootLocation = Paths.get(storagePath);
+		this.rootLocation = Path.of(storagePath);
 		try {
 			Files.createDirectories(rootLocation);
 		} catch (IOException e) {
