@@ -1,5 +1,6 @@
 package com.haust.shop.core.util;
 
+import java.io.Serial;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
@@ -14,7 +15,8 @@ import com.haust.shop.core.exception.UtilException;
  * @author 韩雪磊
  */
 public final class UUID implements java.io.Serializable, Comparable<UUID> {
-	private static final long serialVersionUID = -1185015143654744140L;
+    @Serial
+    private static final long serialVersionUID = -1185015143654744140L;
 
 	/**
 	 * SecureRandom 的单例
