@@ -59,7 +59,7 @@ public class DtsAccountService {
 		List<DtsUserAccount> accounts = userAccountMapper.selectByExample(example);
 		// Assert.state(accounts.size() < 2, "同一个用户存在两个账户");
 		if (accounts.size() == 1) {
-			return accounts.get(0);
+			return accounts.getFirst();
 		} else {
 			logger.error("根据代理用户id：{},获取账号信息出错!!!",shareUserId);
 			return null;
@@ -183,7 +183,7 @@ public class DtsAccountService {
 			conditionSql = "and o.order_status in (" + conditionSql.substring(1) + ") ";
 		}
 		if (settlementStatus != null && settlementStatus.size() == 1) {
-			conditionSql = conditionSql + " and o.settlement_status =" + settlementStatus.get(0) + " ";
+			conditionSql = conditionSql + " and o.settlement_status =" + settlementStatus.getFirst() + " ";
 		}
 
 		PageHelper.startPage(page, size);
