@@ -3,11 +3,13 @@ package com.haust.shop.admin.util;
 
 import com.haust.service.domain.product.DtsBrand;
 
+import java.io.Serial;
 import java.io.Serializable;
 
 public class DtsBrandVo extends DtsBrand implements Serializable{
 
-	private static final long serialVersionUID = 6530090986580196500L;
+    @Serial
+    private static final long serialVersionUID = 6530090986580196500L;
 	
 	private Integer[] categoryIds;
 
