@@ -1,11 +1,13 @@
 package com.haust.shop.admin.domain;
 
+import java.io.Serial;
 import java.io.Serializable;
 import java.math.BigDecimal;
 
 public class AccountVo implements Serializable {
 
-	private static final long serialVersionUID = 1567048369574496965L;
+    @Serial
+    private static final long serialVersionUID = 1567048369574496965L;
 
 	private Integer userId;
 
