@@ -3,12 +3,14 @@ package com.haust.shop.admin.domain;
 
 import com.haust.common.util.CategorySellAmts;
 
+import java.io.Serial;
 import java.io.Serializable;
 import java.util.List;
 
 public class CategorySellVo implements Serializable{
-	
-	private static final long serialVersionUID = 96458407347975166L;
+
+    @Serial
+    private static final long serialVersionUID = 96458407347975166L;
 
     private String[] categoryNames;//一级大类目录名称
 	
