@@ -1,14 +1,17 @@
 package com.haust.common.util;
 
+import java.io.Serial;
 import java.io.Serializable;
 import java.math.BigDecimal;
+
 /** 
 * @Description: 这里进行的是categoryamts的名称和值，amts的目的值控制线程安全的意思
 * @Author: 韩雪磊
 */ 
 public class CategorySellAmts implements Serializable{
 
-	private static final long serialVersionUID = 677901688504280013L;
+    @Serial
+    private static final long serialVersionUID = 677901688504280013L;
 
 	private String name;
 	
