@@ -103,7 +103,7 @@ public class DtsCouponServiceImpl implements DtsCouponService {
 		} else if (couponList.size() == 0) {
 			return null;
 		} else {
-			return couponList.get(0);
+			return couponList.getFirst();
 		}
 	}
 
