@@ -80,7 +80,7 @@ public class DtsCouponUserServiceImpl implements DtsCouponUserService {
 		if (couponUserList.size() == 0) {
 			return null;
 		}
-		return couponUserList.get(0);
+		return couponUserList.getFirst();
 	}
 
 	public DtsCouponUser findById(Integer id) {
