@@ -1,5 +1,6 @@
 package com.haust.service.domain.order;
 
+import java.io.Serial;
 import java.io.Serializable;
 import java.math.BigDecimal;
 import java.util.List;
@@ -12,7 +13,8 @@ import java.util.List;
  */
 public class BrandOrderGoods implements Serializable {
 
-	private static final long serialVersionUID = 4756437344642762485L;
+    @Serial
+    private static final long serialVersionUID = 4756437344642762485L;
 
 	private Integer brandId;
 
