@@ -1,5 +1,6 @@
 package com.haust.shop.wx.dao;
 
+import java.io.Serial;
 import java.io.Serializable;
 import java.math.BigDecimal;
 import java.util.List;
@@ -14,7 +15,8 @@ import com.haust.shop.db.domain.DtsOrderGoods;
  */
 public class BrandOrderGoods implements Serializable {
 
-	private static final long serialVersionUID = 4756437344642762485L;
+    @Serial
+    private static final long serialVersionUID = 4756437344642762485L;
 
 	private Integer brandId;
 
