@@ -1,5 +1,6 @@
 package com.haust.shop.wx.dao;
 
+import java.io.Serial;
 import java.io.Serializable;
 
 /**
@@ -10,7 +11,8 @@ import java.io.Serializable;
  */
 public class UserInfo implements Serializable {
 
-	private static final long serialVersionUID = -5813029516433359765L;
+    @Serial
+    private static final long serialVersionUID = -5813029516433359765L;
 
 	private Integer userId;
 	private String nickName;
